Rename authentication classes in test directory
