+++ budgetr-backend/src/main/java/pl/mantiscrab/budgetr/auth/UserAuthConfig.java
@@ -10,4 +10,9 @@ class UserAuthConfig {
     UserAuthService userAuthService(PasswordEncoder passwordEncoder, UserAuthRepository userAuthRepository) {
         return new UserAuthService(passwordEncoder, userAuthRepository);
     }
+
+    @Bean
+    SecurityContextSignedInUsernameGetter securityContextSignedInUsernameGetter() {
+        return new SecurityContextSignedInUsernameGetter();
+    }
 }
