+++ budgetr-backend/src/test/java/pl/mantiscrab/budgetr/auth/UserAuthTestConfig.java
@@ -2,7 +2,7 @@ package pl.mantiscrab.budgetr.auth;
 
 import org.springframework.security.crypto.password.NoOpPasswordEncoder;
 
-class UserTestConfig {
+class UserAuthTestConfig {
     static UserAuthService userService() {
         return new UserAuthConfig().userAuthService(NoOpPasswordEncoder.getInstance(), new InMemoryUserAuthRepository());
     }
