+++ budgetr-backend/src/test/java/pl/mantiscrab/budgetr/auth/UserAuthTestDataProvider.java
@@ -2,7 +2,7 @@ package pl.mantiscrab.budgetr.auth;
 
 import pl.mantiscrab.budgetr.auth.dto.UserRegisterDto;
 
-class UserTestDataProvider {
+class UserAuthTestDataProvider {
     public static UserRegisterDto.UserRegisterDtoBuilder sampleRegisterDto() {
         return UserRegisterDto.builder()
                 .email("user@user")
